Weverything to Wall

+++ picojson.h
@@ -716,7 +716,7 @@ namespace picojson {
         if (ch == -1 || ch == '\n') {
           break;
         } else if (ch >= ' ') {
-          err->push_back(ch);
+          err->push_back(static_cast<int>(ch));
         }
       }
     }
